+++ CHANGES.md
@@ -23,6 +23,7 @@ Changes in 12.6
 - SQL add csp_reports table in rosariosis.sql & rosariosis_mysql.sql
 - Add Content Security Policy plugin in plugins/Content_Security_Policy/, rosariosis.sql & rosariosis_mysql.sql
 - Update to version 12.6 in Update.fnc.php
+- Add Content-Security-Policy-Report-Only HTTP header in Warehouse.php
 
 Changes in 12.5
 ---------------